fix(SingleResult.vue): Add overflow-x auto to the div surrounding the rendered input url for the SingleResult component.
This covers the extreme cases where super long urls are passed in to be converted that get forced out of their container, and allow the user to scroll horizontally to see all of the link. Nor should the container be expanded any further vertically and forcing wrap of the link.